Add extractDescriptionFragment()
When navigating to a secondary page within the tooltip popup, I
previously used a regex to extract the contents of the <body> tag and
pass it to tippy.setContent(). This was brittle and unwieldly, because
the <body> tag contained stuff like <script> tags.

Instead, extractDescriptionFragment() uses DOMParser to generate a
transient Document object. Then it locates the <div> tag containing the
actual description (and no <script> tags) and copies its contents into
a DocumentFragment. This can be passed to tippy.setContent() without
worrying about <script> tags sneaking in.