fix: Move XPathMatch.ash under release/relay/
Previously, I put this under release/scripts/ because I believed it's a
utility library that can be called from relay and gCLI scripts alike.
However, putting it there causes KoLmafia to add it to the "Scripts".
This menu item does nothing (since XPathMatch.ash does not have main()),
but this is nevertheless confusing and disruptive.

Moving it to release/relay/ causes it to be hidden from KoLmafia's menu.
Eventually, I hope I can phase it out altogether, possibly when I
migrate to JavaScript-driven relay scripts.